Move recursive key removal into it's own package
